pyproject.toml
[project]
name = "20260301-tasklist-agent"
version = "0.1.0"
description = "CLI Task Agent powered by Google Gemini"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alex-leontiev-toolbox-python @ git+https://github.com/nailbiter/alex_leontiev_toolbox_python@main-stable",
    "click>=8.3.2",
    "flask==3.0.0",
    "google-cloud-firestore>=2.19.0",
    "google-genai>=0.3.0",
    "gunicorn==21.2.0",
    "jinja2>=3.1.0",
    "langchain-core>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-firestore>=0.1.7",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.7.0",
    "pymongo[srv]>=4.10.0",
    "python-dotenv>=1.0.1",
    "python-json-logger>=4.1.0",
    "requests>=2.31.0",
]

[tool.black]
# isort -> black
# https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html
[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "langgraph-cli>=0.4.19",
]
